requirements.txt
pygame==2.5.2
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
pygame-gui==0.6.9